✨ separate "plugin_instance_wrapper" to headers & sources
